Move PXIe-5654 driver to separate module
