+++ qcodes_contrib_drivers/drivers/NationalInstruments/RFSG.py
@@ -4,7 +4,6 @@ from typing import Optional, List, Any
 from functools import partial
 import ctypes
 
-from qcodes.utils.validators import Numbers
 from qcodes.utils.helpers import create_on_off_val_mapping as on_off_map
 
 from .visa_types import (
@@ -244,20 +243,3 @@ class NationalInstruments_RFSG(NIDLLInstrument):
 
 # shorthand alias for the above
 NI_RFSG = NationalInstruments_RFSG
-
-
-class PXIe_5654(NI_RFSG):
-    """
-    Device-specific driver for the PXIe-5654 signal generator.
-    """
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        # device-specific parameter limits
-        self.frequency.vals = Numbers(250e3, 20e9)
-        self.power_level.vals = Numbers(-7, 15)
-
-        # check for amplitude extender and update power limits accordingly
-        model = self.IDN()["model"]
-        if "PXIe-5696" in model:
-            self.power_level.vals = Numbers(-110, 24)
